The Voice
- BaseVocals.h Expanded.
Functionality for vocal arrangements increased. As a consequence,
GenArrVocal.cpp/.h have been made redundant and, as such, deleted.

- Vocal Arrangements are in!
Theoretically, at least. Still needs to be tested, of course. Still
need to find a way of identifying the type of track without resorting
to fixed track names. In a distant GUI-rich future, this will be
trivial, but for now it's a concern.