Calm Before the Storm
A fair bit added to the 'write' section, particularly an implementation
for writing notes to the file. However, a serious rethink is needed for
the heirarchy of the arrangement class. Additionally, changes to the
'chord' part of the file require a reworking of all that. All in all, a
ball-ache lies ahead.