Over the Hill
Progress continues towards a working implementation.

- A compromise for how the heirarchy should work is through the use of
a very literal interpretation of pointers.
- Difficulty now has 'pointers' (ints) that refer to the 'ids' of the
Notes/Chords/etc. from a main vector.
- Phrases and Sections are now entirely separate to the Difficulty
tree. A Difficulty can still be sub-divided with getXWithinTime() and
the number of notes will be compared to the previous Difficulty. This
should be enough to prevent either over- or under-correction.

- Commiting here as I intend to start using C++11 features.