It Lives 3: Tokyo Drift
Mostly done now, though testing is still an issue. It's becoming more
and more of a hassle to use the index system rather than pointers, but
I'm reluctant to refactor (read: bugger everything up) again so soon
after getting it to work.

Some more error handling is in, including a debug flag which gets it to
shut up a bit.

Not yet implemented (probably not conclusive):
- Handshapes (Safety stuff in place)
- Techniques (Not convinced it's correctly identifying them)

Further development:
- Refactor for pointers?
- Start working towards extendability: more read/write options. Consider
a tab format or Music XML.